* WIP: refactor backend: group task-specific code inside plugin folders   * DONE serializer mappings     * DONE refactor label_serializer
