+++ backend/src/labelit/serializers/label_serializer.backup.py
@@ -1,0 +1,161 @@
+from rest_framework import serializers
+from labelit.models import (
+    Label,
+    OrdinalLabel,
+    TranscriptionLabel,
+    LiveCorrectLabel,
+    EntityLabel,
+    TextEditionLabel,
+    NestedCategoricalLabel,
+    AudioRegionLabel,
+)
+from rest_polymorphic.serializers import PolymorphicSerializer
+from django.core.exceptions import ObjectDoesNotExist
+from .timed_transcript_serializer import TimedTranscriptSerializer
+
+
+class LabelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Label
+        fields = ["id", "name", "task", "color"]
+
+
+class EntityLabelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = EntityLabel
+        fields = [
+            "id",
+            "name",
+            "task",
+            "color",
+            "start_offset",
+            "end_offset",
+            "source_label",
+        ]
+
+
+class AudioRegionLabelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = AudioRegionLabel
+        fields = [
+            "id",
+            "name",
+            "task",
+            "start",
+            "end",
+        ]
+
+
+class NestedCategoricalLabelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = NestedCategoricalLabel
+        fields = [
+            "id",
+            "name",
+            "task",
+            "parent_label",
+            "single_child_select",
+        ]
+
+
+class OrdinalLabelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = OrdinalLabel
+        fields = [
+            "id",
+            "name",
+            "task",
+            "color",
+            "index",
+        ]
+
+    def create(self, validated_data):
+        try:
+            OrdinalLabel.objects.get(
+                task=validated_data.get("task"), index=validated_data.get("index")
+            )
+            raise serializers.ValidationError(
+                "Validation error: duplicate. A label with matching (task, index) already exists"
+            )
+        except ObjectDoesNotExist:
+            pattern = OrdinalLabel(**validated_data)
+            pattern.save()
+        return pattern
+
+
+class TextEditionLabelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = TextEditionLabel
+        fields = [
+            "id",
+            "name",
+            "task",
+            "color",
+            "edited_text",
+        ]
+
+
+class LiveCorrectLabelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = LiveCorrectLabel
+        fields = [
+            "id",
+            "name",
+            "task",
+            "color",
+            "timed_transcript",
+        ]
+
+
+class LiveCorrectLabelReadSerializer(serializers.ModelSerializer):
+    timed_transcript = TimedTranscriptSerializer(
+        read_only=True,
+    )
+
+    class Meta:
+        model = LiveCorrectLabel
+        fields = [
+            "id",
+            "name",
+            "task",
+            "color",
+            "timed_transcript",
+        ]
+
+
+class TranscriptionLabelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = TranscriptionLabel
+        fields = [
+            "id",
+            "name",
+            "task",
+            "color",
+            "transcript",
+        ]
+
+
+class LabelPolymorphicSerializer(PolymorphicSerializer):
+    model_serializer_mapping = {
+        Label: LabelSerializer,
+        OrdinalLabel: OrdinalLabelSerializer,
+        TranscriptionLabel: TranscriptionLabelSerializer,
+        LiveCorrectLabel: LiveCorrectLabelSerializer,
+        EntityLabel: EntityLabelSerializer,
+        TextEditionLabel: TextEditionLabelSerializer,
+        NestedCategoricalLabel: NestedCategoricalLabelSerializer,
+        AudioRegionLabel: AudioRegionLabelSerializer,
+    }
+
+
+class LabelPolymorphicReadSerializer(PolymorphicSerializer):
+    model_serializer_mapping = {
+        Label: LabelSerializer,
+        OrdinalLabel: OrdinalLabelSerializer,
+        TranscriptionLabel: TranscriptionLabelSerializer,
+        LiveCorrectLabel: LiveCorrectLabelReadSerializer,
+        EntityLabel: EntityLabelSerializer,
+        TextEditionLabel: TextEditionLabelSerializer,
+        NestedCategoricalLabel: NestedCategoricalLabelSerializer,
+        AudioRegionLabel: AudioRegionLabelSerializer,
+    }
